test: add uninstall testing [WIP]

+++ cli/cmd/uninstall.go
@@ -24,11 +24,12 @@ import (
 	k8sutils "github.com/odigos-io/odigos/k8sutils/pkg/client"
 	"go.uber.org/multierr"
 	v1 "k8s.io/api/core/v1"
+	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	k8slabels "k8s.io/apimachinery/pkg/labels"
 	"k8s.io/apimachinery/pkg/selection"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/apimachinery/pkg/util/wait"
-	k8slabels "k8s.io/apimachinery/pkg/labels"
 
 	"github.com/spf13/cobra"
 )
@@ -467,6 +468,10 @@ func uninstallConfigMaps(ctx context.Context, client *kube.Client, ns, _ string)
 	}
 
 	for _, i := range list.Items {
+		// Skip kube-root-ca.crt ConfigMap as it's a system-level resource
+		if i.Name == "kube-root-ca.crt" {
+			continue
+		}
 		err = client.CoreV1().ConfigMaps(ns).Delete(ctx, i.Name, metav1.DeleteOptions{})
 		if err != nil {
 			return err
@@ -480,6 +485,11 @@ func removeAllSources(ctx context.Context, client *kube.Client) error {
 	l := log.Print("Removing Odigos Sources...")
 	sources, err := client.OdigosClient.Sources("").List(ctx, metav1.ListOptions{})
 	if err != nil {
+		// no sources found, nothing to do here
+		if apierrors.IsNotFound(err) || len(sources.Items) == 0 {
+			l.Success()
+			return nil
+		}
 		return err
 	}
 
@@ -492,6 +502,10 @@ func removeAllSources(ctx context.Context, client *kube.Client) error {
 	}
 
 	if deleteErr != nil {
+		if apierrors.IsNotFound(deleteErr) {
+			l.Success()
+			return nil
+		}
 		return deleteErr
 	}
